Reformat with rustfmt.

+++ proptest/src/arbitrary/primitives.rs
@@ -12,16 +12,12 @@
 use crate::bool;
 use crate::char;
 use crate::num::{
-    isize, usize, f32, f64, i16, i32, i64, i8, u16, u32, u64, u8
+    f32, f64, i16, i32, i64, i8, isize, u16, u32, u64, u8, usize,
 };
 #[cfg(not(target_arch = "wasm32"))]
 use crate::num::{i128, u128};
 
-arbitrary!(
-    bool,
-    i8, i16, i32, i64, isize,
-    u8, u16, u32, u64, usize
-);
+arbitrary!(bool, i8, i16, i32, i64, isize, u8, u16, u32, u64, usize);
 
 #[cfg(not(target_arch = "wasm32"))]
 arbitrary!(i128, u128);
